Improve opcode resolution in binary decodification
This commit changes how the opcode is resolved when decoding an instruction.

Instead of checking that a particular field name matches the "opcode" string,
now it looks for a field name which starts with such a string instead. This
allows ISAs to specify multiple opcode fields, which is useful for when an
ISA has multiple formats for an opcode (e.g. opcodes with different lengths).