Improve RISCV Fixes
This commit improves the RISCV fixes in order to accommodate future
expansion.

Instead of having code specific for each fix, a data structure is used
which contains the necessary information in order to code and decode an
instruction which has immediate values split into multiple fields with
a specific codification in each.