Improve ASM RISCV Fixes
Similarly to commit ecd618e, this commit improves asm-related fixes
for the RISCV architecture for better maintainability. The modus
operandi is the same as before: instead of having individual code
lines which perform some fix, indicate all the fixes which need to
be performed in a data structure and have a common piece of code
handle all those cases.