Improve processing of binary regions in Microprobe Test (MPT) files
Before this change, it was assumed that MPT files contained one
instruction per line (in assembly or binary). If the code was in
binary, then it was also assumed that the data would come in the
correct sizes (i.e. the right instruction codification length).

However, this is not the case as Microprobe supports the
decodification of binary streams of instructions of arbitrary
lengths. Splitting the data into the appropiate lengths is
therefore not the responsibility of the program generating the
MPT, but Microprobe's.

This commit changes the code so that the internal representation
of MPT files contains both assembler code and binary code data
in separate data structures to be processed separately (i.e.
by utils/asm.py and utils/bin.py respectively).

Note that, although combining assembly and binary data in the MPT
files is technically supported, the tool for generating C files
out of MPTs won't order those instructions correctly. I am not
sure if this is a common use case or not. If necessary, this
feature would be implemented in a later time.