Fix decodification of long binary data
In earlier commits, support for different endianess of
architectures was added. However, only the parsing of short binary
code regions was tested. The implementation had a bug in which
the whole region would be byte swapped depending on the endianess,
resulting in the reversal of the ordering of the instructions.

This commit reverses some of the changes introduced in f53689a,
and reimplementing them in the correct way.