Compute parameters when decoding single instruction from binary
When the single parameter in `interpret_bin` is true, there will
only be one instruction in the passed binary stream. This means that
the endianess will be "big" (because of how positional notation
works) and that the word length can be computed (using ceil in case
redundant zeros are missing from the input binary number and the
calculated word length is not an integer).