[RISCV] mpt2elf and RISCV overhaul
This large PR implements several improvements and fixes on the RISC-V back-end and to the mpt2elf
tool.  Besides fixes and improvements, new back-ends and environments have been implemented to
target better the SPIKE simulation environment as well as the Sargantana and eprocessor RISC-V platforms. 
The improvements and fixes have been tested during the development of the ChopStiX tool [1] and
allowed the full trace and characterization of NAS, SPEC and other benchmark suites, in the context
of the research collaboration between IBM and BSC (Barcelona Supercomputing Center).

[1] https://github.com/IBM/chopstix/

Signed-off-by: Arnau Bigas <arnau.bigas@gmail.com>
Co-authored-by: Ramon Bertran Monfort <rbertra@us.ibm.com>